@@ -1,29 +1,20 @@
 import { Injectable } from '@angular/core';
-import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HttpResponse } from '@angular/common/http';
-import { Observable, of } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
+import { Observable } from 'rxjs';
 
 import { CachingService } from './caching.service';
 
-@Injectable()
+@Injectable({
+    providedIn: 'root'
+})
 export class CachingInterceptor implements HttpInterceptor {
 
     constructor(private cachingService: CachingService) { }
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-
         switch (req.method) {
-
             case 'GET':
-                const isCached: boolean = this.cachingService.isCached(req.url);
-                return isCached
-                    ? of(this.cachingService.get(req.url))
-                    : next.handle(req).pipe(tap(value => {
-                        if (value instanceof HttpResponse) {
-                            this.cachingService.add(req.url, value);
-                        }
-                    }));
-
+                return this.cachingService.handle(req, next);
             case 'POST':
             case 'PUT':
             case 'DELETE':
