@@ -26,7 +26,7 @@ import { PoultryService } from '../poultry/poultry.service';
   imports: [
     CommonModule,
     CrudModule,
-    RouterModule.forChild([{ path: '', loadChildren: '../crud/crud.module#CrudModule' }]),
+    RouterModule.forChild([{ path: '', loadChildren: () => import('../crud/crud.module').then(m => m.CrudModule) }]),
   ],
 })
 export class PoultryInspectionGroupsModule { }
